Switch to a longpoll-based changes feed to reduce the number of 409 errors.
There's often a decent chunk of time between a worker claiming tasks.
During that time a continuous feed can easily get full of tasks that are
already claimed or even deleted.

Changing to a longpoll approach does mean more HTTP requests but
dramatically reduces the number of 409 Conflict errors when trying to
claim the next task.